[feat & test](定点数版本.测角): 添加radar_dual_channel_clac_angle()用于双通道相位法测角，以及综合测试
- Source/fixed_point/Include/radar_types.h
  Source/fixed_point/Source/radar_doa.c
  Source/fixed_point/Include/radar_doa.h
  使用于双通道雷达相位差测角函数
  - 输入CFAR结果，RDM，以及雷达波长/天线间距
  - 从RDM中查询复数信号，计算相位差，进一步得到角度，然后保存到量测值列表

- Source/fixed_point/Source/radar_processer.c
  Test/Python/fixed_point/test_main.py
  合并测角功能，并可视化结果，测角基本正常

+++ Source/fixed_point/Include/radar_types.h
@@ -12,8 +12,8 @@
 typedef struct {
     int32_t distance;    // 径向距离(mm)
     int32_t velocity;    // 径向速度(mm/s)
-    int32_t azimuth;     // 方位角 (Q2.29)
-    int32_t sin_azimuth; // sin(azimuth) Q0.31
+    int16_t azimuth;     // 方位角 (Q2.13)
+    int16_t sin_azimuth; // sin(azimuth) Q0.15
     int32_t amp;
     int32_t snr;
 } radar_measurements_fixed_t;
